[None][fix] address lucaslie and refine Gemma4 AD multimodal masking
- Move SemanticMaskRegistry to custom_ops/ so models can self-register
  at import time with no circular dependencies. By the time transforms
  run, all model registrations are already in place.
- Inline Gemma4 mask ops and registration into modeling_gemma4.py
  (delete standalone gemma4_multimodal_mask.py).
- Remove get_cached_attention_extra_args from AttentionDescriptor and
  both backends. The KV cache transform passes the prepared mask
  directly — no new abstraction layer needed.
- Simplify _process_semantic_mask to use _add_or_retrieve_input utility
  instead of manual placeholder lookup.
- Fix sliding window position calculation in triton paged context kernel.
- Synchronous H2D copy for extra args to prevent overlap scheduler races.
