+++ tensorrt_llm/_torch/auto_deploy/models/custom/modeling_gemma4.py
@@ -67,6 +67,182 @@ from tensorrt_llm.inputs.registry import (
     MultimodalPlaceholderMetadata,
 )
 
+from ...custom_ops.semantic_mask_registry import SemanticMaskLoweringSpec, SemanticMaskRegistry
+
+# ---------------------------------------------------------------------------
+# Multimodal semantic mask custom ops — prefill-only and cached variants.
+# Defined here so the ops are registered when the module is imported.
+# ---------------------------------------------------------------------------
+
+
+def _blob_ids_from_spans(
+    seq_len: int,
+    positions: torch.Tensor,
+    lengths: torch.Tensor,
+    device: torch.device,
+) -> torch.Tensor:
+    """Build contiguous blob ids from multimodal span positions and lengths."""
+    blob_ids = torch.zeros(seq_len, dtype=torch.int64, device=device)
+    next_blob_id = 1
+    for position, length in zip(positions.tolist(), lengths.tolist(), strict=True):
+        if length <= 0:
+            continue
+        start = max(int(position), 0)
+        end = min(start + int(length), seq_len)
+        if start >= end:
+            continue
+        blob_ids[start:end] = next_blob_id
+        next_blob_id += 1
+    return blob_ids
+
+
+def _causal_or_bidirectional_mask(blob_ids: torch.Tensor) -> torch.Tensor:
+    """Build a bool allow-mask from blob ids."""
+    seq_len = blob_ids.shape[0]
+    causal = torch.tril(torch.ones(seq_len, seq_len, device=blob_ids.device, dtype=torch.bool))
+    q_blob = blob_ids.unsqueeze(1)
+    kv_blob = blob_ids.unsqueeze(0)
+    bidirectional = (q_blob == kv_blob) & (q_blob != 0)
+    return causal | bidirectional
+
+
+@torch.library.custom_op("auto_deploy::gemma4_multimodal_mask", mutates_args=())
+def gemma4_multimodal_mask(
+    input_ids: torch.Tensor,
+    mm_token_positions: torch.Tensor,
+    mm_token_lengths: torch.Tensor,
+    mm_item_cu_seqlen: torch.Tensor,
+    mm_item_types: Optional[torch.Tensor] = None,
+    mm_special_offsets_cu_seqlen: Optional[torch.Tensor] = None,
+    mm_special_offsets: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    """Build a dense source-graph attention mask from multimodal spans."""
+    del mm_item_types, mm_special_offsets_cu_seqlen, mm_special_offsets
+    batch_size, seq_len = input_ids.shape[:2]
+    masks = []
+    for batch_idx in range(batch_size):
+        span_start = int(mm_item_cu_seqlen[batch_idx].item())
+        span_end = int(mm_item_cu_seqlen[batch_idx + 1].item())
+        blob_ids = _blob_ids_from_spans(
+            seq_len,
+            mm_token_positions[span_start:span_end],
+            mm_token_lengths[span_start:span_end],
+            input_ids.device,
+        )
+        masks.append(_causal_or_bidirectional_mask(blob_ids))
+    return torch.stack(masks, dim=0).unsqueeze(1)
+
+
+@gemma4_multimodal_mask.register_fake
+def _gemma4_multimodal_mask_fake(
+    input_ids: torch.Tensor,
+    mm_token_positions: torch.Tensor,
+    mm_token_lengths: torch.Tensor,
+    mm_item_cu_seqlen: torch.Tensor,
+    mm_item_types: Optional[torch.Tensor] = None,
+    mm_special_offsets_cu_seqlen: Optional[torch.Tensor] = None,
+    mm_special_offsets: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    del mm_token_positions, mm_token_lengths, mm_item_cu_seqlen
+    del mm_item_types, mm_special_offsets_cu_seqlen, mm_special_offsets
+    batch_size, seq_len = input_ids.shape[:2]
+    return torch.empty(batch_size, 1, seq_len, seq_len, dtype=torch.bool, device=input_ids.device)
+
+
+@torch.library.custom_op("auto_deploy::gemma4_prepare_multimodal_mask", mutates_args=())
+def gemma4_prepare_multimodal_mask(
+    batch_info_host: torch.Tensor,
+    cu_seqlen: torch.Tensor,
+    input_pos: torch.Tensor,
+    mm_token_positions: torch.Tensor,
+    mm_token_lengths: torch.Tensor,
+    mm_item_cu_seqlen: torch.Tensor,
+    mm_item_types: Optional[torch.Tensor] = None,
+    mm_special_offsets_cu_seqlen: Optional[torch.Tensor] = None,
+    mm_special_offsets: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    """Build backend-consumable prefill masks from multimodal spans and batch geometry."""
+    del mm_item_types, mm_special_offsets_cu_seqlen, mm_special_offsets
+    num_prefill = int(batch_info_host[0].item())
+    if num_prefill == 0 or mm_token_positions.numel() == 0:
+        return torch.empty(0, 1, 0, 0, dtype=torch.bool, device=input_pos.device)
+
+    device = input_pos.device
+    masks = []
+    max_q = 0
+    max_kv = 0
+    for batch_idx in range(num_prefill):
+        q_start = int(input_pos[batch_idx].item())
+        q_len = int(cu_seqlen[batch_idx + 1].item()) - int(cu_seqlen[batch_idx].item())
+        kv_len = q_start + q_len
+
+        span_start = int(mm_item_cu_seqlen[batch_idx].item())
+        span_end = int(mm_item_cu_seqlen[batch_idx + 1].item())
+        blob_ids = _blob_ids_from_spans(
+            kv_len,
+            mm_token_positions[span_start:span_end],
+            mm_token_lengths[span_start:span_end],
+            device,
+        )
+
+        q_blob = blob_ids[q_start : q_start + q_len].unsqueeze(1)
+        kv_blob = blob_ids.unsqueeze(0)
+        bidirectional = (q_blob == kv_blob) & (q_blob != 0)
+
+        q_pos = torch.arange(q_start, q_start + q_len, device=device).unsqueeze(1)
+        kv_pos = torch.arange(kv_len, device=device).unsqueeze(0)
+        causal = kv_pos <= q_pos
+
+        mask = (causal | bidirectional).unsqueeze(0)
+        masks.append(mask)
+        max_q = max(max_q, q_len)
+        max_kv = max(max_kv, kv_len)
+
+    padded = []
+    for mask in masks:
+        _, q_len, kv_len = mask.shape
+        padded.append(F.pad(mask, (0, max_kv - kv_len, 0, max_q - q_len), value=False))
+    return torch.stack(padded, dim=0)
+
+
+@gemma4_prepare_multimodal_mask.register_fake
+def _gemma4_prepare_multimodal_mask_fake(
+    batch_info_host: torch.Tensor,
+    cu_seqlen: torch.Tensor,
+    input_pos: torch.Tensor,
+    mm_token_positions: torch.Tensor,
+    mm_token_lengths: torch.Tensor,
+    mm_item_cu_seqlen: torch.Tensor,
+    mm_item_types: Optional[torch.Tensor] = None,
+    mm_special_offsets_cu_seqlen: Optional[torch.Tensor] = None,
+    mm_special_offsets: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    del mm_token_positions, mm_token_lengths, mm_item_cu_seqlen
+    del mm_item_types, mm_special_offsets_cu_seqlen, mm_special_offsets
+    num_prefill = int(batch_info_host[0].item())
+    q_lens = torch.diff(cu_seqlen[: num_prefill + 1])
+    max_q = int(q_lens.max().item()) if num_prefill > 0 else 0
+    max_kv = int((input_pos[:num_prefill] + q_lens).max().item()) if num_prefill > 0 else 0
+    return torch.empty(num_prefill, 1, max_q, max_kv, dtype=torch.bool, device=input_pos.device)
+
+
+# ---------------------------------------------------------------------------
+# Self-register semantic mask lowering specs.  The registry lives in
+# ``custom_ops/`` which is imported before ``models/``, so this runs at
+# model import time with no circular-dependency issues.
+# ---------------------------------------------------------------------------
+_GEMMA4_MASK_SPEC = SemanticMaskLoweringSpec(
+    prepare_op=torch.ops.auto_deploy.gemma4_prepare_multimodal_mask.default,
+    num_outputs=1,
+)
+SemanticMaskRegistry.register(
+    torch.ops.auto_deploy.gemma4_multimodal_mask, "torch", _GEMMA4_MASK_SPEC
+)
+SemanticMaskRegistry.register(
+    torch.ops.auto_deploy.gemma4_multimodal_mask, "triton_paged", _GEMMA4_MASK_SPEC
+)
+
+
 # ---------------------------------------------------------------------------
 # Bundled config classes — enables loading on transformers <5.3 where
 # Gemma4 is not natively registered.
@@ -1253,9 +1429,6 @@ class Gemma4ForCausalLM(Gemma4TextPreTrainedModel, GenerationMixin):
         input_ids: Optional[torch.LongTensor] = None,
         position_ids: Optional[torch.LongTensor] = None,
         inputs_embeds: Optional[torch.Tensor] = None,
-        batch_info_host: Optional[torch.Tensor] = None,
-        cu_seqlen: Optional[torch.Tensor] = None,
-        input_pos: Optional[torch.Tensor] = None,
         mm_item_cu_seqlen: Optional[torch.Tensor] = None,
         mm_token_positions: Optional[torch.Tensor] = None,
         mm_token_lengths: Optional[torch.Tensor] = None,
@@ -1264,7 +1437,7 @@ class Gemma4ForCausalLM(Gemma4TextPreTrainedModel, GenerationMixin):
         mm_special_offsets: Optional[torch.Tensor] = None,
         **kwargs,
     ) -> Gemma4CausalLMOutput:
-        del batch_info_host, cu_seqlen, input_pos, kwargs
+        del kwargs
         assert position_ids is not None, "position_ids must be provided"
 
         if (input_ids is None) == (inputs_embeds is None):
@@ -2402,6 +2575,9 @@ class Gemma4ADInputProcessor:
         self.image_token_id = image_token_id
         self.boi_token_id = boi_token_id
         self.eoi_token_id = eoi_token_id
+        # Bypass the generic multimodal hashing probe. This processor already
+        # constructs ``multimodal_input`` directly from Gemma4 image spans.
+        self.multimodal_hashing_supported = False
 
     def __getattr__(self, name):
         return getattr(self.base, name)
@@ -2500,10 +2676,6 @@ class Gemma4TextExportInfo(TextModelExportInfo):
 
     def _init_dynamic_shape_lookup(self):
         dynamic_shapes = super()._init_dynamic_shape_lookup()
-        dynamic_shapes["batch_info_host"] = None
-        dynamic_shapes["cu_seqlen"] = {0: Dim.DYNAMIC}
-        dynamic_shapes["input_pos"] = {0: Dim.DYNAMIC}
-        dynamic_shapes["seq_len_with_cache"] = {0: Dim.DYNAMIC}
         dynamic_shapes["mm_item_cu_seqlen"] = {0: Dim.DYNAMIC}
         dynamic_shapes["mm_item_types"] = {0: Dim.DYNAMIC}
         dynamic_shapes["mm_token_positions"] = {0: Dim.DYNAMIC}
