+++ tensorrt_llm/_torch/auto_deploy/transform/__init__.py
@@ -1,7 +1,4 @@
 """AutoDeploy's modular graph transform + inference optimizer pipeline."""
 
-from . import (
-    library,  # noqa: F401 - ensure all transforms are registered
-    semantic_mask_registry,  # noqa: F401
-)
+from . import library  # noqa: F401 - ensure all transforms are registered
 from .interface import *
