+++ tests/unittest/auto_deploy/singlegpu/custom_ops/attention/test_torch_attention_op.py
@@ -584,7 +584,7 @@ class TestTorchBackendAttention:
             sinks,
             sliding_window_size,
             logit_cap,
-            # DYNAMIC INPUTS
+            # OPTIONAL INPUTS
             custom_attn_mask=None,
         )
